Rename Facade to Debugbar Creating the facade as a class named Debugbar enables better auto completion for IDEs and the like. To avoid naming conflicts it exists in a separate folder Facades.
For backwards compatibility of existing code the current Facade persists as a class that extends the Debugbar facade class. This should allow for a minor update to the package with a deprecation notice on existing code importing the Facade explicitly.
