+++ composer.json
@@ -48,7 +48,7 @@
                 "Barryvdh\\Debugbar\\ServiceProvider"
             ],
             "aliases": {
-                "Debugbar": "Barryvdh\\Debugbar\\Facade"
+                "Debugbar": "Barryvdh\\Debugbar\\Facades\\Debugbar"
             }
         }
     },
